works, but

+++ distributedRMI/src/RemoteInterface.java
@@ -4,7 +4,7 @@ import data.ServerMonitor;
 
 public interface RemoteInterface extends Remote {
 	
-	int remoteComputations (ServerMonitor monitor) throws RemoteException;
+	long remoteComputations (ServerMonitor monitor) throws RemoteException;
 	int getCoresNumber () throws RemoteException;
 	void setNodeId (int id, int offset) throws RemoteException;
